GRIM: Move set related lua opcodes to thier own file
